backend/pyproject.toml
[project]
name = "cse485-project"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "groq>=0.32.0",
    "mcp[cli]>=1.15.0",
    "python-dotenv>=1.1.1",
    "starlette>=0.48.0",
    "uvicorn[standard]>=0.37.0",
    "httpx>=0.27.0",
    "PyJWT>=2.8.0",
    "openai>=1.109.1",
    "pathlib>=1.0.1",
    "google-auth-oauthlib>=1.2.2",
    "google-api-python-client>=2.184.0",
    "google-auth-httplib2>=0.2.0",
    "bcrypt>=5.0.0",
    "langchain>=1.0.5",
    "langchain-mcp-adapters>=0.1.12",
    "langchain-openai>=1.0.2",
    "langchain-groq>=1.0.0",
    "cryptography>=46.0.3",
    "telesignenterprise>=3.0.0",
    "setuptools>=80.9.0",
    "boto3>=1.42.55",
]

[tool.setuptools]
packages = ["my_app"]
